feat: add shared library
