+++ userService/index.js
@@ -1,15 +1,15 @@
 const mongoose = require('mongoose');
-const config = require('./src/config/config');
-const {logger} = require('./src/config/logger');
+const {config} = require('../libraries/config/config');
 const http =  require('http')
-const {getUsers} = require('./src/getUsers')
-const {createUser} = require('./src/createUser')
-const {getUser} = require('./src/getUser')
+const {getUsers} = require('./src/services/getUsers')
+const {createUser} = require('./src/services/createUser')
+const {getUser} = require('./src/services/getUser')
 const {errorHandler} = require('../libraries/errors/errors.handler')
 const validations = require('../libraries/validations')
 const {validate} = require('../libraries/utilities/validate')
 const {APIError} =  require('../libraries/utilities/APIError')
 const morgan = require('./src/config/morgan')
+const {routes} = require('./src/routes')
 // const app = require('./src/app')
 const express = require('express')
 
@@ -42,9 +42,7 @@ server.listen(config.port,()=>{
 })
 app.use(morgan.errorHandler)
 app.use(morgan.successHandler)
-app.get('/users',getUsers)
-app.post('/users',validate(validations.createUser),createUser)
-app.get('/users/:id',validate(validations.getUser),getUser)
+app.use(routes)
 // global error handler should come after all other middlewares
 app.all('*',(req,res,next)=>{
   const err =  new APIError(404,`requested resource not found on server: ${req.originalUrl}`)
